Additional screenshots for PLNT 2014 2015 2016
PLNT was a unique case because the data showed all zeroes in 2014. It started up in August of 2015 and continued in 2016. This represented a stock that started trading in 2015 and its opening price was from August 2015. This was shown as zero in 2014 because truly, according to the data, there was no change for PLNT and so zero percentage change.